some fixes in all project

+++ src/main/java/bookapp/dto/order/OrderItemDto.java
@@ -3,6 +3,6 @@ package bookapp.dto.order;
 public record OrderItemDto(
         Long id,
         Long bookId,
-        Long quantity
+        int quantity
 ) {
 }
